Refactor component and test architecture
